refactor: use Nuxt 4 directory structure
