+++ app/components/ConfigManager.vue
