+++ app/components/DeepResearch/NodeDetail.vue
