+++ app/components/DeepResearch/SearchNode.vue
