+++ app/components/LangSwitcher.vue
