+++ app/components/PasswordInput.vue
