+++ app/components/ReasoningAccordion.vue
