+++ i18n.config.ts
@@ -1,6 +1,6 @@
-import en from '~/i18n/en.json'
-import zh from '~/i18n/zh.json'
-import nl from '~/i18n/nl.json'
+import en from '~~/i18n/en.json'
+import zh from '~~/i18n/zh.json'
+import nl from '~~/i18n/nl.json'
 
 export default defineI18nConfig(() => ({
   legacy: false,
