+++ lib/feedback.ts
@@ -4,6 +4,8 @@ import { zodToJsonSchema } from 'zod-to-json-schema'
 
 import { languagePrompt, systemPrompt } from './prompt'
 import { useAiModel } from '~/composables/useAiProvider'
+import { parseStreamingJson, type DeepPartial } from '~~/utils/json'
+import { throwAiError } from '~~/utils/errors'
 
 type PartialFeedback = DeepPartial<z.infer<typeof feedbackTypeSchema>>
