+++ nuxt.config.ts
@@ -62,5 +62,8 @@ export default defineNuxtConfig({
 
   css: ['~/assets/css/main.css'],
   compatibilityDate: '2024-11-01',
+  future: {
+    compatibilityVersion: 4,
+  },
   devtools: { enabled: true },
 })
